correct spelling of OS X
The correct spelling for "OS X" includes a space between "OS" and "X".  This makes sense, because "OS" is (of course) an initialism for operating system, and "X" is a Roman numeral (and is pronounced as "ten", by the way).  It would be odd if Apple had decided to jam them together as though X were the initial of another word.  For those of us using Macs going back to the day when dinosaurs walked the earth, it also continues logically from Mac OS 9: http://toastytech.com/guis/macos9about.png.